backend/routes/image_recognition.py
from flask import Blueprint, request, jsonify
from PIL import Image
import io

image_recognition_bp = Blueprint('image_recognition', __name__)

@image_recognition_bp.route('/upload', methods=['POST'])
def upload_image():
    image_file = request.files['image']
    # Here, you would process the image with a model (not implemented)
    return jsonify({"message": "Image processed successfully", "results": "plant disease detected"}), 200
