requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.10.9
TensorFlow==2.6.0
Pillow==8.4.0
scikit-learn==0.24.2
requests==2.26.0
